added impersonation history to support author deletions of messages via X emoji reactions, other fixes and changes

+++ bot/cogs/impersonation/message_tracker.py
@@ -2,11 +2,12 @@ import discord
 from discord import Message
 from discord.ext import commands
 
-from bot.core.bot import user_config
+from bot.core.bot import impersonation_default
+from bot.core.bot import impersonation_history
 from bot.core.bot import Bot
 from bot.utils.settings import settings
 from bot.utils.logger import logger
-from bot.utils.helpers import is_rp_enabled, send_as_profile, get_channel_id, resolve_message_reference
+from bot.utils.helpers import is_rp_enabled, send_as_profile, get_channel_id
 
 
 # ----------------------------------------------------------------------
@@ -57,7 +58,7 @@ class ImpersonationMessageTracker(commands.Cog):
 
         content = message.content.strip() if message.content else ""
 
-        if not content and not message.attachments:
+        if not content and not message.attachments and not message.stickers:
             # Empty message with no attachments, delete it
             try:
                 await message.delete()
@@ -68,11 +69,11 @@ class ImpersonationMessageTracker(commands.Cog):
                 logger.warn(f"Failed to delete message: {e}")
             return
 
-        # Expect format: `trigger: message content`
-        if ":" not in content or content.startswith("https://"):
+        # Expect format: `trigger: message content` (exclude messages with only links)
+        if ":" not in content or content.lower().startswith(("https://", "http://")):
             # Check if user has a default trigger
             user_id = message.author.id
-            default_trigger = await user_config.get_default_trigger(user_id)
+            default_trigger = await impersonation_default.get(user_id)
 
             if default_trigger:
                 # Prepend default trigger to message
@@ -89,28 +90,25 @@ class ImpersonationMessageTracker(commands.Cog):
                     logger.warn(f"Failed to delete message: {e}")
                 return
 
-        trigger_part, content_part = map(str.strip, content.split(":", 1))
-        if not trigger_part:
+        parts = content.split(":", 1)
+        if len(parts) != 2:
+            # Handle missing trigger or invalid format
             try:
                 await message.delete()
-                logger.debug(
-                    f"Deleted message from {message.author} in {message.channel.id} "
-                    "for empty trigger."
-                )
+                logger.debug(f"Deleted message from {message.author} in {message.channel.id} for invalid format.")
             except Exception as e:
                 logger.warn(f"Failed to delete message: {e}")
             return
+
+        trigger_part, content_part = map(str.strip, parts)
 
         # Prepare attachments as discord.File objects
         files: list["discord.File"] = [
             await attachment.to_file() for attachment in message.attachments
         ] if message.attachments else []
 
-        # get relied message if it exists
-        reply_to = await resolve_message_reference(self.bot, message.reference) if message.reference else None
-
         # Send the message via impersonation profile
-        success = await send_as_profile(
+        message_sent: discord.Message | None = await send_as_profile(
             bot=self.bot,
             profile_trigger=trigger_part,
             user=message.author,
@@ -119,16 +117,23 @@ class ImpersonationMessageTracker(commands.Cog):
             attachments=files,
             send_callback=send_callback,
             rm_thinking_callback=rm_thinking,
-            reply_to=reply_to,
+            stickers=message.stickers if message.stickers else None,
         )
+
+        print(message_sent)
+        print(message_sent.id if message_sent else "No message sent")
+
+        if message_sent:
+            # Track the impersonated message
+            await impersonation_history.add(message.author.id, message_sent.id)
 
         # Delete the original message after sending
         try:
             await message.delete()
-            if success:
+            if message_sent:
                 logger.debug(
                     f"Processed impersonation message from {message.author} "
-                    f"in channel {get_channel_id(message.channel)} with trigger '{trigger_part}': \"{content_part}\""
+                    f"in channel {get_channel_id(message.channel)} with trigger '{trigger_part}' (message id: {message_sent.id}): \"{content_part}\""
                 )
             else:
                 logger.debug(
